feat(미션3): OutputView와 InputView 클래스를 View 클래스로 통합
OutputView 필드와 메소드를 InputView로 결합하고 View 클래스로 이름 변경
