+++ src/main/java/kr/codesquad/View.java
@@ -3,18 +3,31 @@ package kr.codesquad;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.Arrays;
 
-public class InputView {
-    private BufferedReader br;
-    // 최소 2명이며 ,로 값을 검증하는 정규식표현
-    private final String PATTERN = "^[a-zA-Z]{1,5}(,[a-zA-Z]{1,5})+$";
+public class View {
+    private static final String HOW_MUCH_HIGH = "최대 사다리 높이는 몇 개인가요?";
+    private static final String WHO_DOES_PARTICIPATE = "참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)";
+    private static BufferedReader br;
 
-    public InputView() {
+    public View() {
         br = new BufferedReader(new InputStreamReader(System.in));
     }
 
-    public String[] getParticipants() {
+    public static String[] asksParticipate() {
+        System.out.println(WHO_DOES_PARTICIPATE);
+        return getParticipants();
+    }
+
+    public static int asksHeight() {
+        System.out.println(HOW_MUCH_HIGH);
+        return getLadderHeight();
+    }
+
+    public static void printLadder(StringBuilder ladder) {
+        System.out.println(ladder);
+    }
+
+    private static String[] getParticipants() {
         try {
             return splitParticipants(isRightFormat(br.readLine()));
         } catch (IllegalArgumentException | IOException e) {
@@ -23,18 +36,20 @@ public class InputView {
         }
     }
 
-    private String[] splitParticipants(String input) {
+    private static String[] splitParticipants(String input) {
         return input.split(",");
     }
 
-    private String isRightFormat(String input) {
+    private static String isRightFormat(String input) {
+        // 최소 2명이며 ,로 값을 검증하는 정규식표현
+        String PATTERN = "^[a-zA-Z]{1,5}(,[a-zA-Z]{1,5})+$";
         if (!input.matches(PATTERN)) {
             throw new IllegalArgumentException("인원수는 최소 2명이며, ','로 구분해주세요.");
         }
         return input;
     }
 
-    public int getLadderHeight() {
+    public static int getLadderHeight() {
         try {
             return isRightLadderHeight(br.readLine());
         } catch (IllegalArgumentException | IOException e) {
@@ -43,7 +58,7 @@ public class InputView {
         }
     }
 
-    private int isRightLadderHeight(String input) {
+    private static int isRightLadderHeight(String input) {
         int height = Integer.parseInt(input);
         if (height <= 0) {
             throw new IllegalArgumentException("사다리 높이는 1 이상입니다.\n");
